// modules/travel/angular/projects/travel/src/lib/models/travel-types.ts
export enum RequestStatus {
  Draft = 0,
  PendingApproval = 1,
  Returned = 2,
  Rejected = 3,
  Approved = 4,
  AtTravelOffice = 5,
  VisaCheck = 6,
  InsuranceCheck = 7,
  FlightSelection = 8,
  TicketsBooked = 9,
  DocsUploaded = 10,
  PendingEmployeeConfirm = 11,
  Confirmed = 12,
  CalculatingAllowances = 13,
  Completed = 14,
  Cancelled = 15,
}

export const RequestStatusLabels: Record<RequestStatus, string> = {
  [RequestStatus.Draft]: 'RequestStatus.Draft',
  [RequestStatus.PendingApproval]: 'RequestStatus.PendingApproval',
  [RequestStatus.Returned]: 'RequestStatus.Returned',
  [RequestStatus.Rejected]: 'RequestStatus.Rejected',
  [RequestStatus.Approved]: 'RequestStatus.Approved',
  [RequestStatus.AtTravelOffice]: 'RequestStatus.AtTravelOffice',
  [RequestStatus.VisaCheck]: 'RequestStatus.VisaCheck',
  [RequestStatus.InsuranceCheck]: 'RequestStatus.InsuranceCheck',
  [RequestStatus.FlightSelection]: 'RequestStatus.FlightSelection',
  [RequestStatus.TicketsBooked]: 'RequestStatus.TicketsBooked',
  [RequestStatus.DocsUploaded]: 'RequestStatus.DocsUploaded',
  [RequestStatus.PendingEmployeeConfirm]: 'RequestStatus.PendingEmployeeConfirm',
  [RequestStatus.Confirmed]: 'RequestStatus.Confirmed',
  [RequestStatus.CalculatingAllowances]: 'RequestStatus.CalculatingAllowances',
  [RequestStatus.Completed]: 'RequestStatus.Completed',
  [RequestStatus.Cancelled]: 'RequestStatus.Cancelled',
};